Separate spectacle functionality and types
