+++ workspaces/ui-v2/src/store/documentationEdit/thunks.ts
@@ -107,11 +107,11 @@ export const saveDocumentationChanges = createAsyncThunk<
         await spectacle.mutate({
           query: `
           mutation X(
-            $commands: [JSON],
-            $batchCommitId: ID,
-            $commitMessage: String,
-            $clientId: ID,
-            $clientSessionId: ID
+            $commands: [JSON!]!,
+            $batchCommitId: ID!,
+            $commitMessage: String!,
+            $clientId: ID!,
+            $clientSessionId: ID!
           ) {
             applyCommands(
               commands: $commands,
